Added hero power page. Updated Armor formular, added missing lava Spawns.

+++ Application/Warcraft3PowerLevels/Entities/General/IUnit.cs
@@ -23,7 +23,7 @@ namespace Warcraft3PowerLevels.Entities.General
 
         public int Health { get; set; }
         public int Mana { get; set; }
-        public int Armor { get; set; }
+        public double Armor { get; set; }
         public ArmorTypeEnum ArmorType { get; set; }
     }
 }
