+++ Application/Warcraft3PowerLevels/Entities/Humans/Units/SiegeEngine_Air.cs
@@ -20,7 +20,7 @@ namespace Warcraft3PowerLevels.Entities.Humans.Units
 
         public int Health { get; set; } = 700;
         public int Mana { get; set; } = 0;
-        public int Armor { get; set; } = 2;
+        public double Armor { get; set; } = 2;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Fortified;
     }
 }
