+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/Bear_1.cs
@@ -15,7 +15,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
         public int Health { get; set; } = 600;
         public int Mana { get; set; } = 0;
 
-        public int Armor { get; set; } = 0;
+        public double Armor { get; set; } = 0;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Heavy;
 
         public double Attack { get; set; } = 20;
