+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/Bear_3.cs
@@ -16,7 +16,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
         public int Health { get; set; } = 1200;
         public int Mana { get; set; } = 0;
 
-        public int Armor { get; set; } = 2;
+        public double Armor { get; set; } = 2;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Heavy;
 
         public double Attack { get; set; } = 39;
