+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/DoomGuard_Ground.cs
@@ -16,7 +16,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
         public int Health { get; set; } = 1600;
         public int Mana { get; set; } = 500;
 
-        public int Armor { get; set; } = 3;
+        public double Armor { get; set; } = 3;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Heavy;
 
         public double Attack { get; set; } = 38.5;
