+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/Hawk_3.cs
@@ -11,7 +11,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
 
         public int Health { get; set; } = 650;
         public int Mana { get; set; } = 0;
-        public int Armor { get; set; } = 5;
+        public double Armor { get; set; } = 5;
 
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Light;
