+++ Application/Warcraft3PowerLevels/Entities/NightElves/Units/Hippogryph.cs
@@ -20,7 +20,7 @@ namespace Warcraft3PowerLevels.Entities.NightElves.Units
 
         public int Health { get; set; } = 525;
         public int Mana { get; set; } = 0;
-        public int Armor { get; set; } = 0;
+        public double Armor { get; set; } = 0;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Unarmored;
     }
 }
