+++ Application/Warcraft3PowerLevels/Entities/Orcs/Summons/SerpentWard_1.cs
@@ -20,7 +20,7 @@ namespace Warcraft3PowerLevels.Entities.Orcs.Summons
 
         public int Health { get; set; } = 90;
         public int Mana { get; set; } = 0;
-        public int Armor { get; set; } = 0;
+        public double Armor { get; set; } = 0;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Heavy;
     }
 }
